backend/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.38
pydantic==2.9.2
swagger-ui-bundle==0.0.9
segment-anything-py==1.0
torch==2.6.0
torchvision==0.21.0
numpy==1.26.4
opencv-python-headless==4.9.0.80
pillow==11.0.0
matplotlib==3.9.2
timm==0.9.16
reportlab==4.2.0
python-multipart==0.0.9
